+++ Server/Content/User.h
@@ -13,8 +13,8 @@ public:
     string name;
     string id;
     uint32_t userIdentifyId;
-    int32_t locationX;
-    int32_t locationY;
+    float locationX;
+    float locationY;
     uint16_t characterId;
     weak_ptr<class GameSession> ownerSession;
 };
